Se agrega la activity de usuario, se pasa la info connun intent y se agrega el Parcelize
