+++ app/build.gradle
@@ -22,6 +22,10 @@ android {
     }
 }
 
+androidExtensions {
+    experimental = true
+}
+
 dependencies {
     def lifecycle_version = "2.1.0"
