+++ app/src/main/java/com/adrianiglesia/ejerciciopracticoflux/Model/Coordinates.kt
@@ -1,0 +1,14 @@
+package com.adrianiglesia.ejerciciopracticoflux.Model
+
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Coordinates(
+    @SerializedName("latitude")
+    val latitude: String,
+    @SerializedName("longitude")
+    val longitude: String
+) : Parcelable
