+++ app/src/main/java/com/adrianiglesia/ejerciciopracticoflux/Model/Location.kt
@@ -1,0 +1,21 @@
+package com.adrianiglesia.ejerciciopracticoflux.Model
+
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Location(
+    @SerializedName("city")
+    val city: String,
+    @SerializedName("coordinates")
+    val coordinates: Coordinates,
+    @SerializedName("postcode")
+    val postcode: Int,
+    @SerializedName("state")
+    val state: String,
+    @SerializedName("street")
+    val street: String
+
+) : Parcelable
