+++ app/src/main/java/com/adrianiglesia/ejerciciopracticoflux/Model/Login.kt
@@ -1,0 +1,24 @@
+package com.adrianiglesia.ejerciciopracticoflux.Model
+
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Login(
+    @SerializedName("md5")
+    val md5: String,
+    @SerializedName("password")
+    val password: String,
+    @SerializedName("salt")
+    val salt: String,
+    @SerializedName("sha1")
+    val sha1: String,
+    @SerializedName("sha256")
+    val sha256: String,
+    @SerializedName("username")
+    val username: String,
+    @SerializedName("uuid")
+    val uuid: String
+) : Parcelable
