+++ app/src/main/java/com/adrianiglesia/ejerciciopracticoflux/Model/Name.kt
@@ -1,0 +1,16 @@
+package com.adrianiglesia.ejerciciopracticoflux.Model
+
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Name(
+    @SerializedName("first")
+    val first: String,
+    @SerializedName("last")
+    val last: String,
+    @SerializedName("title")
+    val title: String
+) : Parcelable
