+++ app/src/main/java/com/adrianiglesia/ejerciciopracticoflux/Model/Picture.kt
@@ -1,0 +1,16 @@
+package com.adrianiglesia.ejerciciopracticoflux.Model
+
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Picture(
+    @SerializedName("large")
+    val large: String,
+    @SerializedName("medium")
+    val medium: String,
+    @SerializedName("thumbnail")
+    val thumbnail: String
+) : Parcelable
