+++ app/src/main/java/com/adrianiglesia/ejerciciopracticoflux/Model/User.kt
@@ -1,0 +1,30 @@
+package com.adrianiglesia.ejerciciopracticoflux.Model
+
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class User(
+    @SerializedName("cell")
+    val cell: String,
+    @SerializedName("dob")
+    val dob: Dob,
+    @SerializedName("email")
+    val email: String,
+    @SerializedName("gender")
+    val gender: String,
+    @SerializedName("location")
+    val location: Location,
+    @SerializedName("login")
+    val login: Login,
+    @SerializedName("name")
+    val name: Name,
+    @SerializedName("nat")
+    val nat: String,
+    @SerializedName("phone")
+    val phone: String,
+    @SerializedName("picture")
+    val picture: Picture
+) : Parcelable
