+++ app/src/main/res/layout/activity_user.xml
@@ -7,4 +7,50 @@
         android:layout_height="match_parent"
         tools:context=".View.UserActivity">
 
+    <ImageView
+            android:layout_width="0dp"
+            android:layout_height="150dp"
+            app:srcCompat="@mipmap/ic_launcher"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:id="@+id/user_image"
+            app:layout_constraintTop_toTopOf="parent"/>
+    <TextView
+            android:text="Aldana"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp" android:id="@+id/user_first_name"
+            android:layout_marginTop="16dp" app:layout_constraintTop_toBottomOf="@+id/user_image"
+            android:textSize="36sp"
+    />
+    <TextView
+            android:text="Vazquez"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/user_last_name" app:layout_constraintStart_toEndOf="@+id/user_first_name"
+            android:layout_marginStart="16dp" android:layout_marginTop="16dp"
+            app:layout_constraintTop_toBottomOf="@+id/user_image"
+            android:textSize="36sp"/>
+    <TextView
+            android:text="15/03/1995"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
+            android:id="@+id/tv_date"
+            android:layout_marginTop="8dp" app:layout_constraintTop_toBottomOf="@+id/tv_nat"
+            android:textSize="14sp"/>
+    <TextView
+            android:text="AR"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
+            android:id="@+id/tv_nat" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toBottomOf="@+id/user_first_name" android:textSize="18sp"/>
+    <TextView
+            android:text="aldana.vazques@gmail.com"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="16dp"
+            android:id="@+id/user_mail" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toBottomOf="@+id/tv_date" android:textSize="14sp"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
